chore(form): decouple common field functionality into separate withBaseField
