+++ packages/core/src/form/index.ts
@@ -2,3 +2,4 @@ export * from './reatomField'
 export * from './reatomFieldArray'
 export * from './reatomFieldSet'
 export * from './reatomForm'
+export * from './withBaseField'
